Feature/#129 base template asset linking

+++ tasks/generate.js
@@ -4,6 +4,7 @@ const productBuildConfigFileName = 'esds-build-config',
         config = require('./config.js'),
         gulp = config.getGulpInstance(),
         c = config.get(),
+        clear = require('clear'),
         fs = require('fs-extra'),
         path = require('path'),
         inquirer = require('inquirer'),
@@ -38,7 +39,10 @@ function copyDefaultStarterFiles(rootPath) {
         starterFiles = [
             {
                 file: path.join(defaultTemplatesDir, 'docs', 'index.njk'),
-                destination: path.join(rootPath, c.docsPath, `index${c.markupSourceExtension}`)
+                destination: path.join(rootPath, c.docsPath, `index${c.markupSourceExtension}`),
+                replace: {
+                  'TEMPLATES_DIRECTORY': c.templatesPath
+                }
             },
             {
                 file: path.join(defaultTemplatesDir, '.gitignore-default'),
@@ -50,17 +54,44 @@ function copyDefaultStarterFiles(rootPath) {
             },
             {
                 file: path.join(defaultTemplatesDir, 'templates', 'sink.njk'),
-                destination: path.join(rootPath, c.templatesPath, `sink${c.markupSourceExtension}`)
+                destination: path.join(rootPath, c.templatesPath, `sink${c.markupSourceExtension}`),
+                replace: {
+                  'TEMPLATES_DIRECTORY': c.templatesPath
+                }
             },
             {
                 file: path.join(defaultTemplatesDir, 'templates', 'base.njk'),
-                destination: path.join(rootPath, c.templatesPath, `base${c.markupSourceExtension}`)
+                destination: path.join(rootPath, c.templatesPath, `base${c.markupSourceExtension}`),
+                replace: {
+                  'CODE_NAMESPACE': c.codeNamespace
+                }
+            },
+            {
+                file: path.join(defaultTemplatesDir, 'tokens', 'tokens.yaml'),
+                destination: path.join(rootPath, c.tokensPath, `tokens.yaml`)
+            },
+            {
+                file: path.join(defaultTemplatesDir, 'styles', 'style-default.scss'),
+                destination: path.join(rootPath, c.stylesPath, `${c.codeNamespace}${c.stylesSourceExtension}`)
+            },
+            {
+                file: path.join(defaultTemplatesDir, 'components', 'default-concatenated-macro.njk'),
+                destination: path.join(rootPath, c.componentsPath, `${c.codeNamespace}${c.markupSourceExtension}`)
             }
         ];
 
     starterFiles.forEach(sf => {
         if (!fs.existsSync(sf.destination)) {
-            fs.copySync(sf.file, sf.destination);
+            if (sf.replace) {
+              let template = fs.readFileSync(sf.file, 'UTF-8');
+              for (let search in sf.replace) {
+                const regex = new RegExp(search, "g");
+                template = template.replace(regex, sf.replace[search]);
+              }
+              fs.writeFileSync(sf.destination, template, 'UTF-8');
+            } else {
+              fs.copySync(sf.file, sf.destination);
+            }
         }
     });
 }
@@ -128,8 +159,8 @@ function generateComponentFiles(answers, rootPath) {
 {% block body %}
     {% filter markdown %}
         # ${componentSinkTitle} Sink
-    {% endfilter %} 
-    {# Your sink examples go here #} 
+    {% endfilter %}
+    {# Your sink examples go here #}
 {% endblock body %}`;
 
         if (!fs.existsSync(componentSinkDirectory)) {
@@ -155,7 +186,7 @@ gulp.task(config.getBaseTaskName(taskNames.generateNewComponent), function(done)
                 default: false,
                 message: "Generate a javascript file for this component?"
             }];
-
+    clear();
     return inquirer.prompt(questions).then(function(answers) {
         generateComponentFiles(answers, c.rootPath);
         done();
